+++ common/tft/tft.c
@@ -30,6 +30,6 @@ void tft_fill_rectangle(uint16_t x1,uint16_t y1,uint16_t x2,uint16_t y2, uint16_
 	ll_tft_fill_rectangle(x1,y1,x2,y2,color);
 }
 
-void tft_draw_bitmap_unscaled(uint16_t x, uint16_t y, uint16_t width, uint16_t height, const uint8_t* dat) {
+void tft_draw_bitmap_unscaled(uint16_t x, uint16_t y, uint16_t width, uint16_t height, const uint16_t* dat) {
 	ll_tft_draw_bitmap_unscaled(x,y,width,height,dat);
 }
