+++ emulator/qt/ll_tft.cpp
@@ -37,7 +37,7 @@ void ll_tft_fill_rectangle(uint16_t x1,uint16_t y1,uint16_t x2,uint16_t y2, uint
      mainwindow->fill_rectangle(x1,y1,x2,y2,color);
 }
 
-void ll_tft_draw_bitmap_unscaled(uint16_t x, uint16_t y, uint16_t width, uint16_t height, const uint8_t* dat) {
+void ll_tft_draw_bitmap_unscaled(uint16_t x, uint16_t y, uint16_t width, uint16_t height, const uint16_t* dat) {
     mainwindow->draw_bitmap_unscaled(x,y,width,height,dat);
 }
